+++ spec/models/signature_spec.rb
@@ -2,9 +2,7 @@ require 'spec_helper'
 
 describe "Signature" do
 
-  before(:each) do
-    Rails.cache.clear
-  end
+  
 
   it "should create a new instance given a valid attribute" do
     s = FactoryGirl.create(:signature, zip: "11211")